Take a process id as unsigned, and name a Win32 error the same way twice.
Two failures under Windows PowerShell, which is what the pipeline runs the suite
with, and which neither of them reached before because the C# had only ever been
driven from PowerShell 7 here.

`Get-ADTWindowTitle` declared `-ParentProcessId` as UInt32[] while
`WindowInfoOptions` took a ReadOnlyCollection<int>. PowerShell 7 converts that
element by element; Windows PowerShell attempts the covariant array assignment
and throws ArrayTypeMismatchException, so every call naming a parent process id
failed there. A process id is never negative, so the C# moves to unsigned rather
than the parameter to signed: the filter, the id `WindowInfo` reports, and the
window enumeration that already had one in hand as a uint. The identifier now
round trips - what Get-ADTWindowTitle hands out can be passed straight back in -
and `WindowUtilities` reads the id it already resolved instead of asking the
process a second time.

`Get-ADTMsiExitCodeMessage` named exit code 0 differently depending on the
runtime. Code 0 is ERROR_SUCCESS, NO_ERROR and DNS_ERROR_RCODE_NO_ERROR all at
once, and `Enum.ToString()` settles a value carrying several names by an internal
lookup that answers differently on .NET Framework and .NET: an install that
succeeded reported ERROR_SUCCESS under one and DNS_ERROR_RCODE_NO_ERROR under the
other. `ExceptionUtilities` now chooses for itself - the canonical winerror.h
spelling, then the shortest name, then alphabetical order - and its Win32 message
suffix, which carried the same latent problem, goes through the same choice.

`Enum.GetNames<TEnum>()` and `Enum.GetValues<TEnum>()` come from
Meziantou.Polyfill for net472, so the generic overloads CA2263 asks for can be
used on both frameworks.

+++ src/PSADT.ClientServer.Client/ClientExecutable.cs
@@ -335,7 +335,7 @@ namespace PSADT.ClientServer
                                                     // Start gracefully closing each open window.
                                                     foreach (WindowInfo window in windows)
                                                     {
-                                                        Process process = Process.GetProcessById(window.ParentProcessId);
+                                                        Process process = Process.GetProcessById((int)window.ParentProcessId);
                                                         await closeAppsDialogState.LogAction($"Closing window with title [{window.WindowTitle}] for process [{process.ProcessName}], prompting to save if necessary.", LogSeverity.Info).ConfigureAwait(false);
                                                         try
                                                         {
@@ -369,7 +369,7 @@ namespace PSADT.ClientServer
                                                         Stopwatch promptToCloseStopwatch = Stopwatch.StartNew();
                                                         while (true)
                                                         {
-                                                            if (!WindowUtilities.GetProcessWindowInfo(parentProcessIdFilter: [process.Id], windowHandleFilter: [window.WindowHandle]).Any())
+                                                            if (!WindowUtilities.GetProcessWindowInfo(parentProcessIdFilter: [(uint)process.Id], windowHandleFilter: [window.WindowHandle]).Any())
                                                             {
                                                                 await closeAppsDialogState.LogAction($"Window [{window.WindowTitle}] for process [{process.ProcessName}] was successfully closed.", LogSeverity.Info).ConfigureAwait(false);
                                                                 break;
